Set up Atom link parameters even if a unique place name was entered
Some code paths directly rendered the page, whereas we really want
them to drop through to the respond_to block, which adds parameters
for the Atom link.  This meant that, for example, the Atom link
was not shown on pages like:

  http://www.fixmytransport.com/issues/browse?name=Brighton%2C+Brighton+%26+Hove

Thanks to Myfanwy Nixon for pointing out that bug.