Link to the Loughborough survey on questionnaire completion
It seems very unlikely that anyone will attempt to answer
the survey multiple times, but to be on the safe side, we're going
to record a random value in the database for each survey link
generated, so that we can exclude anyone trying to game the
survey before analysis.  We add a parameter of the form:

  c1234-73ebc2109ba31123

... and record such generated links in a new external_questionnaires
table.  (The first letter is 'c' or 'p' for either a campaign or
a problem, while the succeding digits are the ID of that issue.
The letters after the dash are a random string.)