Add option to order issues by creation / modification
This version of a fix for #299 just uses the updated_at
column of problems for the modification date of problems.
This column is updated on adding comments, creation and
confirmation (as we'd want) but is also updated at
any other time the object is saved, so the ordering is
somewhat surprising.