When ordering issues by latest activity, show that date instead
The ordering may look confusing if we still show the creation
date in the list when ordering by latest activity, so show the
date of the latest event / update instead in that case.

Arguably, this should be done by insisting that the partial
is rendered with a flag set in locals instead of using an
instance variable.  However, the shared/campaign and
shared/problem partials are rendered via many different
code paths, and changing each render to add that variable
would be a much more invasive change.