Autocomplete station names when finding train routes
This uses JQuery UI's autocomplete functionality to
suggest completions of train station names from 3 or
more typed letters.

The autocomplete_train_station action first tries to
match the string at the beginning of station names, but
if that produces no results, will return any matches for
that string anywhere in the station name.