If an exactly correct station name is entered, don't disambiguate
Previously, if someone typed "Cambridge Rail Station" into
the From or To boxes on the find_train_route page, they would
still be asked to disambiguate between Cambridge Rail Station
and Cambridge Heath (London) Rail Station.  This change makes
an exact match return only that result.

This is particularly important with the autocompletion - if
someone has picked the name from a list, they shouldn't have
to then disambiguate it afterwards.