Dynamically update the "choose from a list of stops" link
The find_stop map page has a link so that you can pick the
visible stops from a list, but this wasn't being updated
when the map was dragged.  It turns out that many users
switch to the list when they can't see what they want on
a busy map, so it's important that this link is in sync.

Fixes #666