Upgrade the jQuery Form Plugin from 2.43 to 3.21
jQuery Form seems to be (now?) distributed unminified, so
also change to keeping an unminified version in the repository
and minifying via the cached_assets:minify task, as with our
other unminified Javascript.

The reason for this upgrade was to try to fix the error
"Failed" in IE8 when there is a jQuery input selector used
when dealing with submission of the "Create Account" form.
It seems that upgrading jQuery, jQuery Form and jQuery UI to
the latest stable versions is sufficient to fix this problem.