Upgrade jQuery from 1.5.2 to 1.8.3
The reason for this upgrade was to try to fix the error
"Failed" in IE8 when there is a jQuery input selector used
when dealing with submission of the "Create Account" form.
It seems that upgrading jQuery, jQuery Form and jQuery UI to
the latest stable versions is sufficient to fix this problem.