Upgrade jQuery UI from 1.8.13 to 1.9.2
I generated a custom jQuery UI download from:

   http://jqueryui.com/download/

... trying to select the same options as previously for jQuery UI.
Previously, going by the comments in the minified Javascript file,
these were:

 * jQuery UI 1.8.13
 * jQuery UI Widget 1.8.13
 * jQuery UI Position 1.8.13
 * jQuery UI Autocomplete 1.8.13
 * jQuery UI Dialog 1.8.13
 * jQuery UI Tabs 1.8.13
 * jQuery UI Effects 1.8.13
 * jQuery UI Effects Blind 1.8.13
 * jQuery UI Effects Fade 1.8.13

In fact, the selection had to be slightly different due to
new dependencies of some components:

  Dialog now requires
    Draggable
    Resizable
      ... which both require
        Mouse

  Autocomplete now requires
    Menu

  Dialog now requires
    Button

So, the eventual selections were:

  jquery.ui.core.js
  jquery.ui.widget.js
  jquery.ui.mouse.js
  jquery.ui.position.js

  jquery.ui.draggable.js
  jquery.ui.resizable.js

  jquery.ui.autocomplete.js
  jquery.ui.button.js
  jquery.ui.dialog.js
  jquery.ui.menu.js
  jquery.ui.tabs.js

  jquery.ui.effect.js
  jquery.ui.effect-blind.js
  jquery.ui.effect-fade.js

It looks as if the CSS for jQuery UI was custom, rather than one
of the preselected styles, so I haven't changed the CSS.

The reason for this upgrade was to try to fix the error
"Failed" in IE8 when there is a jQuery input selector used
when dealing with submission of the "Create Account" form.
It seems that upgrading jQuery, jQuery Form and jQuery UI to
the latest stable versions is sufficient to fix this problem.